+++ pygbnf/__init__.py
@@ -73,6 +73,7 @@ from .tokens import (
 
 # Grammar helpers
 from .helpers import (
+    T,
     WS,
     any_char,
     between,
@@ -80,6 +81,7 @@ from .helpers import (
     float_number,
     identifier,
     keyword,
+    line,
     number,
     separated_by,
     spaced_comma_list,